Add error

+++ pkg/usecase/user_interactor.go
@@ -1,6 +1,10 @@
 package usecase
 
-import "github.com/itsubaki/mackerel-api/pkg/domain"
+import (
+	"errors"
+
+	"github.com/itsubaki/mackerel-api/pkg/domain"
+)
 
 type UserInteractor struct {
 	UserRepository UserRepository
@@ -11,5 +15,9 @@ func (s *UserInteractor) List() (*domain.Users, error) {
 }
 
 func (s *UserInteractor) Delete(userID string) (*domain.User, error) {
+	if !s.UserRepository.Exists(userID) {
+		return nil, &UserNotFound{Err{errors.New("the <userId> that was designated doesn't belong to the organization")}}
+	}
+
 	return s.UserRepository.Delete(userID)
 }
